refactor : repository에 존재하는 서비스 로직 service 단으로 위치 변경
